Added history chart for total CPU and Memory via clickable cards

+++ src/CpuStats.h
@@ -3,7 +3,7 @@
 
 struct CpuStats
 {
-    double percent = 0.0;
+    float percent = 0.0;
 };
 
 #endif // CPU_STATS_H
